+++ Block/Adminhtml/System/Config/ReadOnlyField.php
@@ -9,6 +9,9 @@ use Magento\Framework\Data\Form\Element\AbstractElement;
 /** Display-only information has no saved value to inherit or restore. */
 class ReadOnlyField extends Field
 {
+    /**
+     * @inheritDoc
+     */
     public function render(AbstractElement $element): string
     {
         $element->unsScope()